MAINT: Move the python-based components of `compute_integrals.cc` to a new file
